+++ vtd3_continuous_action.py
@@ -9,12 +9,14 @@ import time
 from distutils.util import strtobool
 
 import gym
+import matplotlib.pyplot as plt
 import numpy as np
 import pybullet_envs  # noqa
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
+from celluloid import Camera
 from torch.utils.tensorboard import SummaryWriter
 
 
@@ -44,6 +46,8 @@ def parse_args():
         help="the entity (team) of wandb's project")
     parser.add_argument("--capture-video", type=lambda x: bool(strtobool(x)), default=False, nargs="?", const=True,
         help="weather to capture videos of the agent performances (check out `videos` folder)")
+    parser.add_argument("--visualize-timestep-distribution", type=lambda x: bool(strtobool(x)), default=False, nargs="?", const=True,
+        help="weather to visualize the distribution of timesteps of states")
 
     # Algorithm specific arguments
     parser.add_argument("--asyncvec", type=lambda x: bool(strtobool(x)), default=False, nargs="?", const=True,
@@ -66,7 +70,7 @@ def parse_args():
         help="the scale of policy noise")
     parser.add_argument("--exploration-noise", type=float, default=0.2,
         help="the scale of exploration noise")
-    parser.add_argument("--learning-starts", type=int, default=25e3,
+    parser.add_argument("--learning-starts", type=int, default=25000,
         help="timestep to start learning")
     parser.add_argument("--policy-frequency", type=int, default=2,
         help="the frequency of training policy (delayed)")
@@ -79,9 +83,25 @@ def parse_args():
     return args
 
 
+class TimestepStats(gym.Wrapper):
+    def __init__(self, env) -> None:
+        super().__init__(env)
+
+    def step(self, action):
+        obs, reward, done, info = super().step(action)
+        self.t += 1
+        return obs, reward, done, info
+
+    def reset(self, **kwargs):
+        obs = super().reset(**kwargs)
+        self.t = 0
+        return obs
+
+
 def make_env(gym_id, seed, idx, capture_video, run_name):
     def thunk():
         env = gym.make(gym_id)
+        env = TimestepStats(env)
         env = gym.wrappers.RecordEpisodeStatistics(env)
         if capture_video:
             if idx == 0:
@@ -179,6 +199,7 @@ if __name__ == "__main__":
     actions = torch.zeros((args.num_steps, args.num_envs) + envs.single_action_space.shape).to(device)
     rewards = torch.zeros((args.num_steps, args.num_envs)).to(device)
     dones = torch.zeros((args.num_steps, args.num_envs)).to(device)
+    timesteps = torch.zeros((args.num_steps, args.num_envs)).to(device)
 
     # TRY NOT TO MODIFY: start the game
     global_step = 0
@@ -187,6 +208,11 @@ if __name__ == "__main__":
     next_done = torch.zeros(args.num_envs).to(device)
     num_updates = args.total_timesteps // args.batch_size
     num_gradient_updates = 0
+
+    # Visualization:
+    if args.visualize_timestep_distribution:
+        fig, ax = plt.subplots()
+        camera = Camera(fig)
 
     for update in range(1, num_updates + 1):
         # ROLLOUTS
@@ -218,6 +244,7 @@ if __name__ == "__main__":
                 torch.Tensor(next_obs).to(device),
                 torch.Tensor(done).to(device),
             )
+            timesteps[step] = torch.Tensor(np.array([env.t for env in envs.envs]))
 
             for item in info:
                 if "episode" in item.keys():
@@ -225,6 +252,16 @@ if __name__ == "__main__":
                     writer.add_scalar("charts/episodic_return", item["episode"]["r"], global_step)
                     writer.add_scalar("charts/episodic_length", item["episode"]["l"], global_step)
                     break
+
+        if args.visualize_timestep_distribution:
+            max_timestep = int(timesteps.max().long())
+            histogram = timesteps.histc(bins=max_timestep, min=0, max=max_timestep)
+            x = range(max_timestep)
+            ax.bar(x, histogram.cpu())
+            ax.set_xlabel("Visited timestep of the obs (higher the better)")
+            ax.set_ylabel("Number of occurrences")
+            ax.text(0.2, 1.01, f"global_step={global_step}, update={update}", transform=ax.transAxes)
+            camera.snap()
 
         # TRAINING
         b_obs = obs.reshape((-1,) + envs.single_observation_space.shape)
@@ -294,6 +331,12 @@ if __name__ == "__main__":
                         target_param.data.copy_(args.tau * param.data + (1 - args.tau) * target_param.data)
         print("SPS:", int(global_step / (time.time() - start_time)))
         writer.add_scalar("charts/SPS", int(global_step / (time.time() - start_time)), global_step)
+
+    if args.visualize_timestep_distribution:
+        animation = camera.animate()
+        animation.save(f"runs/{run_name}/animation.mp4")
+        if args.track:
+            wandb.log({"video.0": wandb.Video(f"runs/{run_name}/animation.mp4")})
 
     envs.close()
     writer.close()
